Change implementation of auth key cache check
If one must be certain of the non-interactivity of
the script, one should not pass the `-a` option,
which allows the authentications keys to be tested
with `ssh-add`, which would cause a `pinentry`
prompt when the wrong password was cached.

The new implementation of the authentication key
check still relies on `ssh-add`, but is now
non-interactive, rendering the `-a` option
useless.